!U Odin Serializer 2020.7.23 - 69b61965 -> 2020.8.12 - 128a2758

+++ CabinIcarus.OdinSerializer/Core/Serializers/ByteSerializer.cs
@@ -36,7 +36,7 @@ namespace CabinIcarus.OdinSerializer
             string name;
             var entry = reader.PeekEntry(out name);
 
-            if (entry == EntryType.Byte)
+            if (entry == EntryType.Integer)
             {
                 byte value;
                 if (reader.ReadByte(out value) == false)
