+++ CabinIcarus.OdinSerializer/Core/Serializers/EnumSerializer.cs
@@ -16,8 +16,6 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-using UnityEngine;
-
 namespace CabinIcarus.OdinSerializer
 {
     using System;
@@ -54,7 +52,8 @@ namespace CabinIcarus.OdinSerializer
         {
             string name;
             var entry = reader.PeekEntry(out name);
-            if (entry == EntryType.ULong)
+
+            if (entry == EntryType.Integer)
             {
                 ulong value;
                 if (reader.ReadUInt64(out value) == false)
